Users now have tasks

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { CreateTaskDTO } from './dto/create-task.dto';
 import { SearchByTaskDTO } from './dto/search-by-task .dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { TaskRepository } from './task.repository';
+import { User } from '../auth/user.entity';
 
 
 @Injectable()
@@ -13,12 +14,12 @@ export class TasksService {
     constructor(@InjectRepository(TaskRepository) private taskRepository: TaskRepository,) {}
 
 
-    async getTask(searchByTask: SearchByTaskDTO): Promise<Task[]> {
-        return this.taskRepository.getTasks(searchByTask);
+    async getTask(searchByTask: SearchByTaskDTO, user: User): Promise<Task[]> {
+        return this.taskRepository.getTasks(searchByTask, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {        
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {        
+        const found = await this.taskRepository.findOne({where: {id, userId: user.id}});
 
         if(!found){
             throw new NotFoundException(`Task with id of "${id}" was not found!`);
@@ -26,21 +27,21 @@ export class TasksService {
         return found
     } 
 
-    async deleteTaskById(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTaskById(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({id, userId: user.id});
         if(result.affected === 0) {
             throw new NotFoundException(`Task with id of "${id}" was not found!`);
         }
     }
 
-    async patchTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const taskToBePatched = await this.getTaskById(id);     
+    async patchTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const taskToBePatched = await this.getTaskById(id, user);     
         taskToBePatched.status = status;
         await taskToBePatched.save();
         return taskToBePatched;
     }
 
-    createTask(createTaskDTO: CreateTaskDTO): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDTO);
+    createTask(createTaskDTO: CreateTaskDTO, user: User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDTO, user);
     }
 }
